feat(service): add catalog + order HTTP routes (auth-gated POST)
Two endpoints on the gtk/v1 surface:

  GET  /api/gtk/v1/services       UNAUTHENTICATED. Active catalog with
       PublicService DTOs (price + category + included_credits, no
       agent_slug, no system_prompt). Marketing/services page reads it.

  POST /api/gtk/v1/service/order  AUTHENTICATED. Creates a
       gtk_service_order row in status='pending_payment', returns the
       new order_no. checkout_url is empty in v0.9 — v0.10 will fill
       with LemonSqueezy hosted checkout (overseas) or hupijiao QR
       (mainland). The auth.RequireAuth path mirrors newapi/router.go.

Storage layer (service/store.go):
- LoadActiveServices: ORDER BY display_order ASC, id ASC for stable
  output. Hydrates PriceDisplayCNY at the boundary.
- LoadServiceBySlug: returns ErrServiceNotFound for missing OR
  inactive — both look the same to a customer.
- NewOrderNo: 8-hex-char random "SVC-AB12CD34" via crypto/rand,
  defended by UNIQUE (order_no) at the table level.
- CreateOrder: snapshots service_slug + price + included_credits at
  order time so retroactive catalog edits don't mutate history.

CoaiUserID promoted to int64 to match auth.User.GetID return type and
keep the call site cast-free.

Build green: go build ./..., go test ./service/ all pass.

+++ service/types.go
@@ -55,7 +55,7 @@ type Service struct {
 type ServiceOrder struct {
 	ID                int64      `json:"id"`
 	OrderNo           string     `json:"order_no"`
-	CoaiUserID        int        `json:"coai_user_id"`
+	CoaiUserID        int64      `json:"coai_user_id"`
 	ServiceID         int64      `json:"service_id"`
 	ServiceSlug       string     `json:"service_slug"`
 	PriceCNYCentsPaid int64      `json:"price_cny_cents_paid"`
